Create FTP server
Date: 02.05.2023

This is an example Python code for creating an FTP server on a Mac OS using the ftplib and socketserver modules. 

In this example, the FTP server listens on port 2121 and serves files from the directory /Users/Shared/ftp. You can change these values to suit your needs.

Note that this example creates a basic, unsecured FTP server. For production use, you should consider adding encryption and user authentication to ensure the security of your files.