Create 005_Doubly-Linked_List_Python_v2.py
In this code, the Node class represents a node in the doubly-linked list. Each node has a data attribute to store the data, a prev attribute to point to the previous node, and a next attribute to point to the next node.

The DoublyLinkedList class represents the doubly-linked list data structure. It has a head attribute, which points to the first node in the list, and a tail attribute, which points to the last node in the list.

The methods implemented in the DoublyLinkedList class are as follows:

is_empty: Checks if the list is empty.
append: Appends an element to the end of the list.
prepend: Prepends an element to the beginning of the list